+++ packages/s3-media/src/bootstrap.ts
@@ -3,9 +3,12 @@ import "reflect-metadata";
 import { Server } from "net";
 import Application from "koa";
 import dependency from "@cheeket/koa";
+import { setGracefulCleanup } from "tmp-promise";
 
 import DependencyInitializer from "./dependency.initializer";
 import router from "./router";
+
+setGracefulCleanup();
 
 async function bootstrap(port?: number): Promise<Server> {
   const application = new Application();
