+++ packages/s3-media/src/service/fs/stat.ts
