+++ packages/s3-media/src/service/tmp/tmpPath.ts
@@ -1,6 +1,4 @@
-import { tmpName, setGracefulCleanup } from "tmp-promise";
-
-setGracefulCleanup();
+import { tmpName } from "tmp-promise";
 
 async function tmpPath(): Promise<string> {
   return tmpName({ prefix: "media" });
